requirements.txt
httpx
chainlit
pandas
plotly
python-dotenv
sqlite3
openai
logging
asyncio
google-generativeai